+++ src/bika/lims/content/abstractanalysis.py
@@ -452,8 +452,7 @@ class AbstractAnalysis(AbstractBaseAnalysis):
         account the Detection Limits.
         :param value: is expected to be a string.
         """
-        # Always update ResultCapture date when this field is modified
-        self.setResultCaptureDate(DateTime())
+        prev_result = self.getField("Result").get(self) or ""
 
         # Convert to list ff the analysis has result options set with multi
         if self.getResultOptions() and "multi" in self.getResultOptionsType():
@@ -498,6 +497,12 @@ class AbstractAnalysis(AbstractBaseAnalysis):
                     else:
                         val = self.getUpperDetectionLimit()
 
+        # Update ResultCapture date if necessary
+        if not val:
+            self.setResultCaptureDate(None)
+        elif prev_result != val:
+            self.setResultCaptureDate(DateTime())
+
         # Set the result field
         self.getField("Result").set(self, val)
 
@@ -523,17 +528,24 @@ class AbstractAnalysis(AbstractBaseAnalysis):
 
         # Add interims to mapping
         for i in interims:
-            if 'keyword' not in i:
+
+            interim_keyword = i.get("keyword")
+            if not interim_keyword:
                 continue
+
             # skip unset values
-            if i['value'] == '':
+            interim_value = i.get("value", "")
+            if interim_value == "":
                 continue
-            try:
-                ivalue = float(i['value'])
-                mapping[i['keyword']] = ivalue
-            except (TypeError, ValueError):
-                # Interim not float, abort
+
+            # Only floatable and UIDs are supported
+            if api.is_floatable(interim_value):
+                interim_value = float(interim_value)
+
+            elif not api.is_uid(interim_value):
                 return False
+
+            mapping[interim_keyword] = interim_value
 
         # Add dependencies results to mapping
         dependencies = self.getDependencies()
@@ -1022,15 +1034,6 @@ class AbstractAnalysis(AbstractBaseAnalysis):
         parent = self.aq_parent
         if parent:
             return parent.absolute_url_path()
-
-    @security.public
-    def getParentTitle(self):
-        """This method is used to populate catalog values
-        This function returns the analysis' parent Title
-        """
-        parent = self.aq_parent
-        if parent:
-            return parent.Title()
 
     @security.public
     def getWorksheetUID(self):