+++ src/senaite/core/browser/static/js/bika.lims.utils.attachments.js
@@ -39,39 +39,6 @@
           $('#addButton').prop('disabled', true);
         }
       });
-      $('#Analysis').combogrid({
-        colModel: [
-          {
-            'columnName': 'analysis_uid',
-            'hidden': true
-          }, {
-            'columnName': 'slot',
-            'width': '10',
-            'label': _t('Slot')
-          }, {
-            'columnName': 'service',
-            'width': '35',
-            'label': _t('Service')
-          }, {
-            'columnName': 'parent',
-            'width': '35',
-            'label': _t('Parent')
-          }, {
-            'columnName': 'type',
-            'width': '20',
-            'label': _t('Type')
-          }
-        ],
-        url: window.location.href.replace('/manage_results', '') + '/attachAnalyses?_authenticator=' + $('input[name="_authenticator"]').val(),
-        showOn: true,
-        width: '650px',
-        select: function(event, ui) {
-          $('#Analysis').val(ui.item.service + (" (slot " + ui.item.slot + ")"));
-          $('#analysis_uid').val(ui.item.analysis_uid);
-          $(this).change();
-          return false;
-        }
-      });
     };
   };