+++ src/senaite/core/upgrade/v02_02_000.py
@@ -18,7 +18,10 @@
 # Copyright 2018-2021 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
+from bika.lims import api
+from bika.lims.interfaces.analysis import IRequestAnalysis
 from senaite.core import logger
+from senaite.core.catalog import ANALYSIS_CATALOG
 from senaite.core.config import PROJECTNAME as product
 from senaite.core.upgrade import upgradestep
 from senaite.core.upgrade.utils import UpgradeUtils
@@ -43,6 +46,80 @@ def upgrade(tool):
 
     # -------- ADD YOUR STUFF BELOW --------
     setup.runImportStepFromProfile(profile, "viewlets")
+    setup.runImportStepFromProfile(profile, "rolemap")
+    setup.runImportStepFromProfile(profile, "workflow")
+
+    # Setup the permission for the edition of analysis conditions
+    setup_edit_analysis_conditions(portal)
+
+    # Preserve all information from service conditions in Sample
+    update_analysis_conditions(portal)
+
 
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
+
+
+def setup_edit_analysis_conditions(portal):
+    """Updates role mappings for analyses that are not yet verified for the
+    permission "Edit Analysis Conditions" to become effective
+    """
+    logger.info("Updating role mappings for Analyses ...")
+    wf_tool = api.get_tool("portal_workflow")
+    workflow = wf_tool.getWorkflowById("senaite_analysis_workflow")
+    query = {
+        "portal_type": "Analysis",
+        "review_state": ["registered", "unassigned", "assigned"]
+    }
+    brains = api.search(query, ANALYSIS_CATALOG)
+    total = len(brains)
+    for num, brain in enumerate(brains):
+        if num and num % 100 == 0:
+            logger.info("Updating role mappings: {0}/{1}".format(num, total))
+        obj = api.get_object(brain)
+        workflow.updateRoleMappingsFor(obj)
+        obj.reindexObject(idxs=["allowedRolesAndUsers"])
+
+    logger.info("Updating role mappings for Analyses [DONE]")
+
+
+def update_analysis_conditions(portal):
+    """Walks through all analyses awaiting for result and sets all information
+    from Service Conditions to them
+    """
+    logger.info("Updating service conditions for Analyses ...")
+    query = {
+        "portal_type": "Analysis",
+        "review_state": ["registered", "unassigned", "assigned", "to_be_verified"]
+    }
+    brains = api.search(query, ANALYSIS_CATALOG)
+    total = len(brains)
+    for num, brain in enumerate(brains):
+        if num and num % 100 == 0:
+            logger.info("Updating conditions: {0}/{1}".format(num, total))
+
+        analysis = api.get_object(brain)
+        if not IRequestAnalysis.providedBy(analysis):
+            continue
+
+        # Extract the original conditions set at service level
+        service = analysis.getAnalysisService()
+        service_conditions = service.getConditions()
+        service_conditions = dict(map(lambda cond: (cond.get("title"), cond),
+                                      service_conditions))
+
+        # Extract the conditions from the analysis and update them accordingly
+        conditions = analysis.getConditions()
+        for condition in conditions:
+            title = condition.get("title")
+            orig_condition = service_conditions.get(title, {})
+            condition.update({
+                "required": orig_condition.get("required", ""),
+                "choices": orig_condition.get("choices", ""),
+                "description": orig_condition.get("description", ""),
+            })
+
+        # Reset the conditions
+        analysis.setConditions(conditions)
+
+    logger.info("Updating service conditions for Analyses [DONE]")